wxt: fontscale and fontsize handling was inconsistent
1) font size was for unknown reasons limited to integer values
2) font scale applied at the level of enhanced code tracking,
   so the "noenhanced" attribute disabled the scaling
3) font scale was applied in two separate places and unapplied
   in another.  This could lead to different bits of text being
   scaled by 0, 1, or even 2 applications of the scale factor.

Bug #2373

+++ src/wxterminal/wxt_term.h
@@ -121,7 +121,7 @@ extern int wxt_ctrl;
 extern int wxt_toggle;
 extern int wxt_redraw;
 extern char *wxt_set_fontname;
-extern int wxt_set_fontsize;
+extern double wxt_set_fontsize;
 extern double wxt_set_fontscale;
 extern t_linecap wxt_linecap;
 extern char wxt_title[MAX_ID_LEN + 1];
